tools/ci/requirements.txt
GitPython
PyGithub
python-gitlab